:recycle: move it into server package
now can do actual dependency injection
