+++ server/src/main/java/amosproj/linter/server/linter/CheckBasics.java
@@ -1,4 +1,4 @@
-package amosproj.linter.crawler;
+package amosproj.linter.server.linter;
 
 import java.io.IOException;
 import java.net.HttpURLConnection;
